#Task3 MovieExporter replaced by MovieExporterJob
